+++ src/test/java/org/ject/support/external/dynamodb/config/DynamoDbInitializerTest.java
@@ -2,7 +2,7 @@ package org.ject.support.external.dynamodb.config;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.ject.support.domain.tempapply.TemporaryApplication;
+import org.ject.support.domain.tempapply.domain.TemporaryApplication;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
