+++ src/test/java/org/ject/support/external/dynamodb/repository/TemporaryApplicationRepositoryTest.java
@@ -6,7 +6,8 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import org.ject.support.domain.tempapply.TemporaryApplication;
+import org.ject.support.domain.tempapply.domain.TemporaryApplication;
+import org.ject.support.domain.tempapply.repository.TemporaryApplicationRepository;
 import org.ject.support.external.dynamodb.domain.CompositeKey;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Order;
